Separate voting procedures and proposals
