+++ eras/conway/test-suite/cardano-ledger-conway-test.cabal
@@ -42,7 +42,7 @@ library
         cardano-ledger-babbage-test,
         cardano-ledger-binary:{cardano-ledger-binary, testlib} >=1.0,
         cardano-strict-containers,
-        cardano-ledger-conway:{cardano-ledger-conway, testlib} >=1.3 && <1.6,
+        cardano-ledger-conway:{cardano-ledger-conway, testlib} ^>=1.6,
         cardano-ledger-core:{cardano-ledger-core, testlib} >=1.3 && <1.5,
         cardano-ledger-allegra ^>=1.2,
         cardano-ledger-mary ^>=1.3,
