+++ libs/cardano-ledger-pretty/cardano-ledger-pretty.cabal
@@ -1,6 +1,6 @@
 cabal-version:      3.0
 name:               cardano-ledger-pretty
-version:            1.2.1.1
+version:            1.2.1.2
 license:            Apache-2.0
 maintainer:         operations@iohk.io
 author:             IOHK
@@ -39,7 +39,7 @@ library
         cardano-ledger-alonzo >=1.2,
         cardano-ledger-babbage >=1.1,
         cardano-ledger-byron,
-        cardano-ledger-conway ^>=1.5,
+        cardano-ledger-conway ^>=1.6,
         cardano-ledger-core >=1.3 && <1.5,
         cardano-ledger-mary >=1.0,
         cardano-ledger-shelley ^>=1.4,
